requirements.txt
viam-sdk==0.45.2
chat-service-api @ git+https://github.com/viam-labs/chat-service-api.git@v0.1.11
openai==1.76.0